Based on PR feedback, extract execution search function to helper
