+++ example/spec/helpers/workflowUtils.js
@@ -1,5 +1,6 @@
 'use strict';
 
+const get = require('lodash.get');
 const unset = require('lodash.unset');
 
 const { loadYmlConfigFile, saveYmlConfigFile } = require('./configUtils.js');
@@ -54,7 +55,42 @@ function removeTaskFromWorkflow(workflowName, taskName, workflowConfigFile) {
   saveYmlConfigFile(config, workflowConfigFile);
 }
 
+function isReingestExecution(taskInput) {
+  return get(
+    taskInput,
+    'cumulus_meta.cumulus_context.reingestGranule',
+    false
+  );
+}
+
+function isExecutionForGranuleId(taskInput, granuleId) {
+  const executionGranuleId = get(
+    taskInput,
+    'payload.granules[0].granuleId'
+  );
+
+  return executionGranuleId === granuleId;
+}
+
+/**
+ * Given a Cumulus Message and a granuleId, test if the message is a re-ingest
+ * of the granule.
+ *
+ * This is used as the `findExecutionFn` parameter of the
+ * `waitForTestExecutionStart` function.
+ *
+ * @param {Object} taskInput - a full Cumulus Message
+ * @param {Object} findExecutionFnParams
+ * @param {string} findExecutionFnParams.granuleId
+ * @returns {boolean}
+ */
+function isReingestExecutionForGranuleId(taskInput, { granuleId }) {
+  return isReingestExecution(taskInput) &&
+    isExecutionForGranuleId(taskInput, granuleId);
+}
+
 module.exports = {
+  isReingestExecutionForGranuleId,
   removeWorkflow,
   removeTaskFromWorkflow
 };
